# webserver_attendence/swagger-api.yaml
openapi: 3.0.0
info:
  title: School Attendance Management System API
  description: |
    학교 출석 관리 시스템 API
    
    **3가지 사용자 역할:**
    - Admin: 시스템 관리자 (학과/학기/과목 관리, 시스템 설정, 리포트)
    - Instructor: 교원 (강의 일정, 출석 관리, 공결 승인, 알림 발송)
    - Student: 수강생 (출석 체크, 공결 신청, 메시지)
    
    **핵심 기능:**
    1. 수업/강의 관리
    2. 출석 관리 (2가지 이상 방식)
    3. 공결/사유 처리
    4. 정정/이의제기
    5. 알림/커뮤니케이션
    6. 감사/보안 로그
    
  version: 1.0.0
  contact:
    name: Development Team
    email: dev@school.ac.kr
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
  - url: https://api.school.ac.kr/v1
    description: Production server
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_token
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 200
        password_hash:
          type: string
          maxLength: 255
        role:
          type: string
          enum: [admin, instructor, student]
        department_id:
          type: integer
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        last_login_at:
          type: string
          format: date-time
          nullable: true
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/User'
        message:
          type: string

    # ========== Department ==========
...
    Course:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
          maxLength: 200
        code:
          type: string
          maxLength: 50
        department_id:
          type: integer
        semester_id:
          type: integer
        instructor_id:
          type: integer
        grade:
          type: integer
          description: "학년 (1~4)"
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CourseInput:
      type: object
      required:
        - title
        - code
        - department_id
        - semester_id
        - instructor_id
        - grade
      properties:
        title:
          type: string
        code:
          type: string
        department_id:
          type: integer
        semester_id:
          type: integer
        instructor_id:
          type: integer
        grade:
          type: integer
          minimum: 1
          maximum: 4
        is_active:
          type: boolean
          default: true

    # ========== Class Session ==========
...
    AttendanceMethod:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          maxLength: 50
          example: "전자출결"
        description:
          type: string
          maxLength: 255
    AttendanceStatusCode:
      type: object
      properties:
        code:
          type: integer
          enum: [0, 1, 2, 3, 4]
          description: "0:미정, 1:출석, 2:지각, 3:결석, 4:공결"
        name:
          type: string
          maxLength: 50
    Attendance:
      type: object
      properties:
        id:
          type: integer
        session_id:
          type: integer
        student_id:
          type: integer
        status_code:
          type: integer
          enum: [0, 1, 2, 3, 4]
        method_id:
          type: integer
          nullable: true
        checked_at:
          type: string
          format: date-time
          nullable: true
        instructor_id:
          type: integer
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        is_changed:
          type: boolean
          description: "정정 여부"
    AttendanceCheckRequest:
      type: object
      required:
        - session_id
      properties:
        session_id:
          type: integer
        verification_code:
          type: string
          description: "인증번호 방식인 경우"
    AttendanceUpdateRequest:
      type: object
      required:
        - status_code
      properties:
        status_code:
          type: integer
          enum: [0, 1, 2, 3, 4]
        reason:
          type: string
          description: "변경 사유 (정정 시)"

    # ========== Excuse Request ==========
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        error:
          type: string
        message:
          type: string
...
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: 로그인
      description: 이메일과 비밀번호로 로그인하고 JWT 토큰 및 쿠키를 발급받습니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 로그인 성공
          headers:
            Set-Cookie:
              schema:
                type: string
                example: session_token=abcde12345; Path=/; HttpOnly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: 로그아웃
      security:
        - cookieAuth: []
      responses:
        '200':
          description: 로그아웃 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /auth/me:
    get:
      tags:
        - Authentication
      summary: 현재 로그인한 사용자 정보
      security:
        - cookieAuth: []
      responses:
        '200':
          description: 사용자 정보
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'

  # ========== Users ==========
...
  /courses:
    get:
      tags:
        - Courses
      summary: 과목 목록 조회
      security:
        - cookieAuth: []
      parameters:
        - in: query
          name: semester_id
          schema:
            type: integer
        - in: query
          name: department_id
          schema:
            type: integer
        - in: query
          name: instructor_id
          schema:
            type: integer
        - in: query
          name: grade
          schema:
            type: integer
      responses:
        '200':
          description: 과목 목록
          content:
            application/json:
              schema:
                type: object
                properties:
                  courses:
                    type: array
                    items:
                      $ref: '#/components/schemas/Course'

    post:
      tags:
        - Courses
      summary: 과목 생성 (Admin/Instructor)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseInput'
      responses:
        '201':
          description: 과목 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /courses/{id}:
    get:
      tags:
        - Courses
      summary: 과목 상세 조회
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 과목 정보
          content:
            application/json:
              schema:
                type: object
                properties:
                  course:
                    $ref: '#/components/schemas/Course'

    put:
      tags:
        - Courses
      summary: 과목 수정 (Admin/Instructor)
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseInput'
      responses:
        '200':
          description: 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

    delete:
      tags:
        - Courses
      summary: 과목 삭제 (Admin)
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 삭제 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # ========== Class Sessions ==========
...
  /attendance/methods:
    get:
      tags:
        - Attendance
      summary: 출석 방식 목록 조회
      security:
        - cookieAuth: []
      responses:
        '200':
          description: 출석 방식 목록
          content:
            application/json:
              schema:
                type: object
                properties:
                  methods:
                    type: array
                    items:
                      $ref: '#/components/schemas/AttendanceMethod'
  /attendance/status-codes:
    get:
      tags:
        - Attendance
      summary: 출석 상태 코드 목록
      security:
        - cookieAuth: []
      responses:
        '200':
          description: 출석 상태 코드
          content:
            application/json:
              schema:
                type: object
                properties:
                  codes:
                    type: array
                    items:
                      $ref: '#/components/schemas/AttendanceStatusCode'
...
  /attendance/check:
    post:
      tags:
        - Attendance
      summary: 출석 체크 (Student)
      description: 학생이 출석을 체크합니다. 전자출결 또는 인증번호 방식.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttendanceCheckRequest'
      responses:
        '200':
          description: 출석 체크 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /attendance/{id}:
    get:
      tags:
        - Attendance
      summary: 출석 기록 상세 조회
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 출석 기록
          content:
            application/json:
              schema:
                type: object
                properties:
                  attendance:
                    $ref: '#/components/schemas/Attendance'

    put:
      tags:
        - Attendance
      summary: 출석 상태 수정 (Instructor - 정정)
      description: 교원이 출석 상태를 정정합니다. 변경 이력이 감사 로그에 기록됩니다.
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttendanceUpdateRequest'
      responses:
        '200':
          description: 출석 정정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
security:
  - cookieAuth: []